fix the test nic pair issue

+++ tools/scripts/test_nicpair.py
@@ -1,5 +1,7 @@
+import select
 import subprocess
 import sys
+import time
 import os
     
 nic_1 = sys.argv[1]
@@ -28,24 +30,37 @@ args = [
     ]
 whole_command = [command] + args
 env=os.environ
-new_env = {k: v for k, v in env.items() if"MPI" not in k}
+new_env = {k: v for k, v in env.items() if "MPI" not in k and "PMIX" not in k and "NCCL" not in k}
+# with open(f'output_{nic_1}.txt', 'w') as f:
+#     for k, v in new_env.items():
+#         f.write(f'{k}={v}\n')
 process = subprocess.Popen(whole_command, shell=False, env=new_env, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
 print(whole_command)
 
-output = ""
-try:
-    while process.poll() is None:
+poll_obj = select.poll()
+poll_obj.register(process.stdout, select.POLLIN)
+
+timeout = 10000
+start_time = time.time()
+max_duration = 60
+
+while True:
+    if time.time() - start_time > max_duration:  
+        print("Execution Time Exceeded")
+        process.terminate() 
+        break
+    
+    poll_result = poll_obj.poll(timeout)
+    if poll_result:
         line = process.stdout.readline().strip()
         if line:
-            output += line + "\n"
             print(line)  # Print output dynamically
             if "No device found" in line:
                 process.terminate()
-                print(pair_key)  # Return the failed pair key
-    process.wait()
-    if process.returncode != 0:
-        print("Subprocess returned non-zero exit code:", process.returncode)
-except subprocess.TimeoutExpired:
-    process.terminate()
-    print("Timeout occurred for NIC pair: ", nic_1, nic_2)
-    print(pair_key)  # Return the failed pair key
+        else:
+            print("Execution Complete")
+            break
+    else:
+        print("Execution Timeout")
+        process.terminate()
+        break
